CVAQATESTWEB/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
playwright==1.49.1
openai==1.6.1
openpyxl==3.1.2
python-multipart==0.0.6
pydantic==2.5.3
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
Pillow==10.1.0
bytez
